+++ neon_drive/__init__.py
@@ -1,3 +1,3 @@
 """Neon Drive."""
 
-__version__ = "5.6.0-beta.1"
+__version__ = "5.6.0-beta.2"